dashboard/requirements.txt
dash==2.14.2
dash-leaflet==1.0.11
dash-table==5.0.0
pandas==2.0.3
plotly==5.9.0
requests==2.31.0
SQLAlchemy==1.4.39
psycopg2-binary==2.9.6